Move 'bcl_to_fastq_info function' from utils.py to bcl2fastq_utils.py (issue #9).
