+++ auto_process_ngs/auto_processor.py
@@ -15,7 +15,6 @@
 
 import sys
 import os
-import re
 import subprocess
 import logging
 import shutil
@@ -1046,7 +1045,7 @@ class AutoProcess:
         if nprocessors is None:
             nprocessors = self.settings.bcl2fastq.nprocessors
         # Collect and store info on underlying bcl2fastq software
-        self.metadata['bcl2fastq_software'] = utils.bcl_to_fastq_info()
+        self.metadata['bcl2fastq_software'] = bcl2fastq_utils.bcl_to_fastq_info()
         # Create bcl2fastq directory
         bcl2fastq_dir = self.add_directory(self.params.unaligned_dir)
         # Get info about the run
